backend/requirements.txt
alembic==1.13.3
awscli==1.36.40
awscli-local==0.22.0
beautifulsoup4==4.13.3
bidict==0.23.1
billiard==4.2.1
blinker==1.8.2
boto3==1.35.42
botocore==1.35.99
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.2
cachelib==0.13.0
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6 
cryptography==44.0.2
Deprecated==1.2.18
dill==0.3.6
dnslib==0.9.26
dnspython==2.7.0
docutils==0.16
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==36.2.2
Flask==3.0.3
Flask-Cors==3.0.10
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-OAuthlib==0.9.6
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
limits==4.0.1
localstack==4.2.0
localstack-client==2.7
localstack-core==4.2.0
localstack-ext==4.2.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
msgpack==1.1.0
oauthlib==2.1.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
plux==1.12.1
prompt_toolkit==3.0.48
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1>=0.4.6, <0.5.0
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyotp==2.9.0
pyparsing==3.2.1
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.4.0
python-socketio==5.12.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==1.1.0
rich==13.9.4
rsa>=3.1.2,<4.8
s3transfer==0.10.3
selenium==4.27.1
semver==3.0.4
sentry-sdk==2.19.2
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
tabulate==0.9.0
tailer==0.4.1
sentry-sdk==2.19.2
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3>=1.25.4,<1.27
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.1.2
flask_socketio
twilio
pytest
eventlet
mixpanel==4.10.1
geopandas
shapely
osmnx